+++ test/usecases/run/UpdateRunUseCase.test.js
@@ -21,7 +21,8 @@ const { expect } = chai;
 
 module.exports = () => {
     const wrongId = 9999999999;
-
+    const TIMESTAMP = 1664271988000;
+    const BIG_INT_NUMBER = '99999999999999999';
     let getRunDto;
     let updateRunDto;
     let updateRunByRunNumberDto;
@@ -53,6 +54,14 @@ module.exports = () => {
                 aliceDipoleCurrent: 45654.1,
                 aliceDipolePolarity: 'NEGATIVE',
                 fillNumber: 1,
+                startOfDataTransfer: TIMESTAMP,
+                endOfDataTransfer: TIMESTAMP,
+                ctfFileCount: 30,
+                ctfFileSize: BIG_INT_NUMBER,
+                tfFileCount: 1234,
+                tfFileSize: BIG_INT_NUMBER,
+                otherFileCount: 123156132,
+                otherFileSize: BIG_INT_NUMBER,
             },
         });
     });
@@ -242,6 +251,14 @@ module.exports = () => {
             expect(result.aliceDipoleCurrent).to.equal(45654.1);
             expect(result.aliceDipolePolarity).to.equal('NEGATIVE');
             expect(result.fillNumber).to.equal(1);
+            expect(result.startOfDataTransfer).to.equal(TIMESTAMP);
+            expect(result.endOfDataTransfer).to.equal(TIMESTAMP);
+            expect(result.ctfFileCount).to.equal(30);
+            expect(result.ctfFileSize).to.equal(BIG_INT_NUMBER);
+            expect(result.tfFileCount).to.equal(1234);
+            expect(result.tfFileSize).to.equal(BIG_INT_NUMBER);
+            expect(result.otherFileCount).to.equal(123156132);
+            expect(result.otherFileSize).to.equal(BIG_INT_NUMBER);
         });
 
         it('Should give an error when the id of the run can not be found', async () => {