+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this project will be documented in this file. See [standa
 
 ## [0.42.0](https://github.com/AliceO2Group/Bookkeeping/releases/tag/%40aliceo2%2Fbookkeeping%400.42.0)
 * Notable changes for users:
+  * Added exta run fields for ccdb in the run details
+* Notable changes for develops:
+  * Runs API:
+    * `GET`
+      * `startOfDataTransfer`, `endOfDataTransfer`, `ctf, tf, other` file size/count are added to the run object.
+    * `PATCH`
+      * `startOfDataTransfer`, `endOfDataTransfer`, `ctf, tf, other` file size/count are added to update run by run number.
   * Added a loading icon when log overview is loading
   * Fixed infinite requests when moving to homepage from logs/runs overview
   * Fixed Trigger Value displayed as active filter if any filter is active on run overview page
   * Run start/stop on the home screen is now consistent with the run overview page.
   * Add CALIBRATION run definition
+  * Pagination limits is now of 1000 items in GUI and 5k in the API and export
 * Notable changes for developers:
   * Some entities and sequelize models have typedefs to help autocompletion, code navigation and refactoring
   * Some UseCases have been reworked to be more reusable and their implementation has been extracted into services (their signature did not changed)